post_notifs.py: skip files that are missing
Newer versions of Pywikibot will throw an exception if you attempt to
fetch content of nonexistent pages. Here we first check if the file
exists, but also use page.text in state.py to avoid the exception being
raised, in the event there are race conditions.

Also fix reference to pywikibot.exceptions.NoPage which is now
pywikibot.exceptions.NoPageError

Bug: T298567