make-list: Only open a ToolsDB connection when it's being used
Instead of opening a single database connection at the start of the
script and then relying that it's still alive after querying MediaWiki
for the affected pages, open the database connection only when starting
to update the database and then nicely close it afterwards (and then
re-open a new one when updating the next category). This should both
avoid holding open unnecessary connections (to comply with the ToolsDB
connection handling policy), and solve the timeouts seen on T339145.

Bug: T339145